AI CodeFusion/requirement.txt
pip install flask requests python-dotenv
Flask==3.0.3
google-generativeai==0.8.3
python-dotenv==1.0.1